+++ src/index.ts
@@ -13,7 +13,7 @@ export * as CurrentPageNumberMiddleware from "./middlewares/CurrentPageNumberMid
 export * as ForceSslMiddleware from "./middlewares/ForceSslMiddleware.js";
 export * as LogRequestMiddleware from "./middlewares/LogRequestMiddleware.js";
 export * as RouterMiddleware from "./middlewares/RouterMiddleware.js";
-export * as StaticMiddleware from "./middlewares/static.js";
+export * as StaticMiddleware from "./middlewares/StaticMiddleware.js";
 
 export * from "./types/HTTPMethod.js";
 export * from "./types/MiddlewareFunction.js";