+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * refactor: made ForceSslMiddleware return its options
 * refactor: made LogRequestMiddleware return its options
 * refactor: made RouterMiddleware return its options
+* refactor: made StaticMiddleware return its options
 
 ## 2.1.0